Toshiki/requirements.txt
git+https://github.com/huggingface/transformers.git
bitsandbytes
transformers
peft
accelerate
trl datasets
sentencepiece
wandb
mecab-python3
unidic-lite
konlpy
unsloth
torchvision
python-mecab-ko
flash-attn